Remove finitization of StringTuple.
This class did more than basic finitization and its performance benefits are a
little illusory. It also assumed the finite versions would all have the same
tuple arity, but that's not at all necessarily the case. And the implementation
was clunky.

This finitization module really represented a fixed-n-ary string tuple. This
would be a valuable feature, and should be added as such, eventually.