Ensure all packs are locked to codeql/util 2.0.0 for best testing.
In the future, we should consider matrix tests to ensure we're compatible with
the ranges that we declare -- both in package versions, and against various CLI
versions.

For the meantime, the best approach is to test the oldest versions, and assume
forward compatibility. At least, that's what's best for codeql/util, if not for
the language packs where our support spans multiple major revisions.