Upgrade to 0.0.3, correct for codeql-util breaking dense rank change. (#15)
* Upgrade to 0.0.3, correct for codeql-util breaking dense rank change.

Previous publication didn't properly test modules with the latest codeql/util,
thus missing a breaking change to `DenseRank` that affected some modules.

Publish a fix to this as 0.0.3 -- fixes the codeql/util constraint, and adapts
modules to the new `DenseRank` API.

* Ensure all packs are locked to codeql/util 2.0.0 for best testing.

In the future, we should consider matrix tests to ensure we're compatible with
the ranges that we declare -- both in package versions, and against various CLI
versions.

For the meantime, the best approach is to test the oldest versions, and assume
forward compatibility. At least, that's what's best for codeql/util, if not for
the language packs where our support spans multiple major revisions.